requirements/base.txt
pytz==2019.3  # https://github.com/stub42/pytz
python-slugify==3.0.6  # https://github.com/un33k/python-slugify
Pillow==6.2.0  # https://github.com/python-pillow/Pillow
argon2-cffi==19.1.0  # https://github.com/hynek/argon2_cffi
whitenoise==4.1.4  # https://github.com/evansd/whitenoise
redis==3.3.11  # https://github.com/antirez/redis
postgis==1.0.4
# Django
# ------------------------------------------------------------------------------
django==2.2.6  # pyup: < 3.0  # https://www.djangoproject.com/
django-environ==0.4.5  # https://github.com/joke2k/django-environ
django-model-utils==3.2.0  # https://github.com/jazzband/django-model-utils
django-redis==4.10.0  # https://github.com/niwinz/django-redis
djangorestframework-gis==0.14
drf-yasg[validation]==1.17

# Django REST Framework
djangorestframework==3.10.3  # https://github.com/encode/django-rest-framework
coreapi==2.3.3  # https://github.com/core-api/python-client
